chore(bedrock): rename custom-resource-provider
